Add figures to paper
Added a number of figures to the paper.  In particular, figures were added to
show the quantile loss (check loss) function, figures showing the phi and lambda
functions, a figure explaining the empirical quantile classifier, and figures
showing the decision rule boundary line for 2-d cases with Gaussian and
exponentiated Gaussian data.